+++ apps/mobile_interface/lib/src/features/solo_practice/controllers/solo_practice_controller.dart
@@ -1,0 +1,122 @@
+import 'package:flutter/material.dart';
+
+import '../models/pronunciation_feedback.dart';
+import '../services/pronunciation_feedback_service.dart';
+
+/// Temporary list of practice prompts used until real lesson data is wired in.
+/// TODO: Replace with a real data source (e.g. API response or lesson model).
+const List<String> _defaultCards = [
+  'The quick brown fox jumped over the lazy dog.',
+  'She sells seashells by the seashore.',
+  'How much wood would a woodchuck chuck?',
+  'Peter Piper picked a peck of pickled peppers.',
+  'I scream, you scream, we all scream for ice cream.',
+  'Red lorry, yellow lorry.',
+  'Unique New York, unique New York.',
+  'Buffalo buffalo Buffalo buffalo buffalo buffalo Buffalo buffalo.',
+  'The sixth sick sheikh\'s sixth sheep\'s sick.',
+  'Fresh French fried fish fingers.',
+  'I saw Susie sitting in a shoeshine shop.',
+  'Lesser leather never weathered wetter weather better.',
+  'Can you can a can as a canner can can a can?',
+  'Willy\'s real rear wheel.',
+  'The thirty-three thieves thought that they thrilled the throne.',
+  'Six sleek swans swam swiftly southwards.',
+  'How can a clam cram in a clean cream can?',
+  'Fuzzy Wuzzy was a bear. Fuzzy Wuzzy had no hair.',
+  'Near an ear, a nearer ear, a nearly eerie ear.',
+  'You\'ve done it! Great work completing the lesson.',
+];
+
+/// Holds session state and logic for the solo practice flow:
+/// current card index, mic state (idle/recording/playback), and feedback.
+class SoloPracticeController {
+  SoloPracticeController({List<String>? cards}) : _cards = cards ?? _defaultCards;
+
+  final List<String> _cards;
+
+  int currentCardIndex = 0;
+
+  /// 0 = idle, 1 = recording, 2 = playback.
+  int micStateIndex = 0;
+
+  PronunciationFeedbackMock? currentFeedback;
+
+  int get totalCards => _cards.length;
+
+  String get currentCard => _cards[currentCardIndex];
+
+  /// Progress value between 0.0 and 1.0 for the progress bar.
+  double get progress => (currentCardIndex + 1) / totalCards;
+
+  /// True when Retry and Submit buttons should be shown (playback state, no feedback yet).
+  bool get showRetrySubmit => micStateIndex == 2 && currentFeedback == null;
+
+  /// True when there is a next card after the current one.
+  bool get hasNextCard => currentCardIndex < totalCards - 1;
+
+  /// Icon for the mic button based on [micStateIndex]: mic, record dot, or play arrow.
+  IconData get currentMicIcon {
+    switch (micStateIndex) {
+      case 1:
+        return Icons.fiber_manual_record;
+      case 2:
+        return Icons.play_arrow;
+      case 0:
+      default:
+        return Icons.mic;
+    }
+  }
+
+  /// Advance mic state: idle → recording → playback. No-op when already in playback.
+  void advanceMicState() {
+    if (micStateIndex < 2) {
+      micStateIndex += 1;
+    }
+  }
+
+  /// Reset mic to idle so the user can re-record.
+  void retry() {
+    micStateIndex = 0;
+  }
+
+  /// Set the current feedback (after submit). Pass null to clear.
+  void setFeedback(PronunciationFeedbackMock? value) {
+    currentFeedback = value;
+  }
+
+  /// Call pronunciation/assess with [audioBytes] and [referenceText], then set
+  /// [currentFeedback] (real result or mock on failure). Does not hold [BuildContext]
+  /// or show dialogs.
+  Future<void> submit({
+    required List<int> audioBytes,
+    required String referenceText,
+    String? accessToken,
+  }) async {
+    final feedback = await fetchPronunciationFeedback(
+      audioBytes: audioBytes,
+      referenceText: referenceText,
+      accessToken: accessToken,
+    );
+    currentFeedback = feedback ?? getMockFeedback(referenceText);
+  }
+
+  /// Clear feedback and advance to next card (or stay on last). Resets mic to idle.
+  /// Returns true if there is a next card, false if this was the last card (caller may show completion dialog).
+  bool advanceToNextCard() {
+    currentFeedback = null;
+    micStateIndex = 0;
+    if (currentCardIndex < totalCards - 1) {
+      currentCardIndex += 1;
+      return true;
+    }
+    return false;
+  }
+
+  /// Reset session to first card, clear feedback, mic to idle (e.g. back button).
+  void resetSession() {
+    currentCardIndex = 0;
+    micStateIndex = 0;
+    currentFeedback = null;
+  }
+}
